Cardano.Benchmarking.Compiler

Documentation

data CompileError where

Constructors

SomeCompilerError :: String -> CompileError

type Compiler a = RWST NixServiceOptions (DList Action) Int (Except CompileError) a

throwCompileError :: CompileError -> Compiler a

maxOutputsPerTx :: Int

type SrcWallet = String

type DstWallet = String

compileOptions :: NixServiceOptions -> Either CompileError [Action]

runCompiler :: NixServiceOptions -> Compiler () -> Either CompileError [Action]

testCompiler :: NixServiceOptions -> Compiler a -> Either CompileError (a, Int, [Action])

compileToScript :: Compiler ()

initConstants :: Compiler ()

importGenesisFunds :: Compiler String

addCollaterals :: SrcWallet -> Compiler (Maybe String)

splittingPhase :: SrcWallet -> Compiler DstWallet

data Split

Constructors

SplitWithChange Coin Int

FullSplits Int

splitSequenceWalletNames :: SrcWallet -> DstWallet -> [Split] -> Compiler [(SrcWallet, DstWallet, Split)]

unfoldSplitSequence :: Coin -> Coin -> Int -> [Split]

benchmarkingPhase :: String -> Maybe String -> Compiler String

data Fees

Constructors

Fees

Fields

_safeCollateral :: Coin

_minValuePerInput :: Coin

evilFeeMagic :: Compiler Fees

emit :: Action -> Compiler ()

logMsg :: Text -> Compiler ()

cmd1 :: (v -> Action) -> (NixServiceOptions -> v) -> Compiler ()

askNixOption :: (NixServiceOptions -> v) -> Compiler v

delay :: Compiler ()

isPlutusType :: TxGenPlutusType -> Compiler Bool

isAnyPlutusMode :: Compiler Bool

newIdentifier :: String -> Compiler String

newWallet :: String -> Compiler String

parseKey :: ByteString -> SigningKey PaymentKey

keyNameGenesisInputFund :: String

keyNameTxGenFunds :: String

keyTxGenFunds :: SigningKey PaymentKey

The key that is used for the very first transaction, i.e. the secure Genesis transaction. addr_test1vzd3muund27y5nw83vymqj3a83pcuzkkejej6s75e5lfjcc85nc3p is the actual address (in Testnet 42). It is also used as change addresse in the first splitting-step.

keyNameSplitPhase :: String

keySplitPhase :: SigningKey PaymentKey

UTxOs that are generated in intermediate splitting steps use: addr_test1vz45dtkyzk6s3245qw8hmaddaatcx8td3pvmntl8ty7q99c22eahm

keyNameBenchmarkInputs :: String

UTxOs of the final splitting steps, i.e. the inputs of the benchmarking phase, use: addr_test1vzj7zv9msmdasvy5nc9jhnn2gqvrvu33v5rlg332zdfrkugklxkau (Plutus script addresses are ofc different.)

keyBenchmarkInputs :: SigningKey PaymentKey

keyNameBenchmarkDone :: String

keyBenchmarkDone :: SigningKey PaymentKey

The output of the actual benchmarking transactions use: addr_test1vz4qz2ayucp7xvnthrx93uhha7e04gvxttpnuq4e6mx2n5gzfw23z Query the progress of the benchmarking phase: `cardano-node query utxo --testnet-magic 42 --address addr_test1vz4qz2ayucp7xvnthrx93uhha7e04gvxttpnuq4e6mx2n5gzfw23z`

keyNameCollaterals :: String

keyCollaterals :: SigningKey PaymentKey

Collateral inputs for Plutus transactions: addr_test1vpckd9muw3l4f8ne4uzumy28p0k84rvx48q46kssjkta5ng4v6sfs