Cardano.Benchmarking.LogTypes

Documentation

data AsyncBenchmarkControl

Constructors

AsyncBenchmarkControl

Fields

abcFeeder :: Async ()

The thread to feed transactions, also called a throttler.

abcWorkers :: [Async ()]

The per-node transaction submission threads.

abcSummary :: IO SubmissionSummary

IO action to emit a summary.

abcShutdown :: IO ()

IO action to shut down the feeder thread.

data BenchTracers

Constructors

BenchTracers

Fields

btTxSubmit_ :: Trace IO (TraceBenchTxSubmit TxId)

btConnect_ :: Trace IO SendRecvConnect

btSubmission2_ :: Trace IO SendRecvTxSubmission2

btN2N_ :: Trace IO NodeToNodeSubmissionTrace

data EnvConsts

Constructors

EnvConsts

Fields

envIOManager :: IOManager

envThreads :: TVar (Maybe AsyncBenchmarkControl)

The reference needs to be a constant, but the referred-to data (AsyncBenchmarkControl) needs to be able to be initialized. This could in principle be an IORef instead of a TVar.

envNixSvcOpts :: Maybe NixServiceOptions

There are situations NixServiceOptions won't be available and defaults will have to be used.

benchTracers :: TVar (Maybe BenchTracers)

This also needs to be accessible to the signal handlers.

data NodeToNodeSubmissionTrace

Constructors

ReqIdsBlocking Ack Req

IdsListBlocking Int

ReqIdsNonBlocking Ack Req

IdsListNonBlocking Int

ReqTxs Int

TxList Int

EndOfProtocol

type SendRecvConnect = WithBearer RemoteConnectionId (TraceSendRecv (Handshake NodeToNodeVersion Term))

type SendRecvTxSubmission2 = TraceSendRecv (TxSubmission2 (GenTxId CardanoBlock) (GenTx CardanoBlock))

data SubmissionSummary

Constructors

SubmissionSummary

Fields

ssTxSent :: !Sent

ssTxUnavailable :: !Unav

ssElapsed :: !NominalDiffTime

ssEffectiveTps :: !TPSRate

ssThreadwiseTps :: ![TPSRate]

ssFailures :: ![String]

data TraceBenchTxSubmit txid

Constructors

TraceTxGeneratorVersion Version

TraceBenchTxSubRecv [txid]
Received from generator.

TraceBenchTxSubStart [txid]
The txid has been submitted to TxSubmission2 protocol peer.

SubmissionClientReplyTxIds [txid]
Announcing txids in response for server's request.

TraceBenchTxSubServReq [txid]
Request for tx received from TxSubmission2 protocol peer.

SubmissionClientDiscardAcknowledged [txid]
An ack (window moved over) received for these transactions.

TraceBenchTxSubServDrop [txid]
Transactions the server implicitly dropped.

SubmissionClientUnAcked [txid]
Transactions outstanding.

TraceBenchTxSubServUnav [txid]
Transactions requested, but unavailable in the outstanding set.

TraceBenchTxSubServFed [txid] Int
Transactions fed by the feeder, accompanied by sequence number.

TraceBenchTxSubServCons [txid]
Transactions consumed by a submitter.

TraceBenchTxSubIdle
Remote peer requested new transactions but none were available, generator not keeping up?

TraceBenchTxSubRateLimit DiffTime
Rate limiter bit, this much delay inserted to keep within configured rate.

TraceBenchTxSubSummary SubmissionSummary
SubmissionSummary.

TraceBenchTxSubDebug String

TraceBenchTxSubError Text

TraceBenchPlutusBudgetSummary PlutusBudgetSummary
PlutusBudgetSummary.

TraceBenchForwardingInterrupted HowToConnect String