Cardano.Benchmarking.OuroborosImports

Documentation

type CardanoBlock = CardanoBlock StandardCrypto

type LocalSubmitTx = TxInMode -> IO TxSubmitResult

data PaymentKey

Shelley-era payment keys. Used for Shelley payment addresses and witnessing transactions that spend from these addresses.

This is a type level tag, used with other interfaces like Key.

data ShelleyGenesis

Shelley genesis information

Note that this is needed only for a pure Shelley network, hence it being defined here rather than in its own module. In mainnet, Shelley will transition naturally from Byron, and thus will never have its own genesis information.

data family SigningKey keyrole

The type of cryptographic signing key, for each key role.

type SigningKeyFile = File (SigningKey ())

data NetworkId

makeLocalConnectInfo :: NetworkId -> SocketPath -> LocalNodeConnectInfo

protocolToTopLevelConfig :: SomeConsensusProtocol -> IO (TopLevelConfig CardanoBlock)

protocolToNetworkId :: SomeConsensusProtocol -> IO NetworkId

protocolToCodecConfig :: SomeConsensusProtocol -> IO (CodecConfig CardanoBlock)

submitTxToNodeLocal :: (HasCallStack, MonadIO m) => LocalNodeConnectInfo -> TxInMode -> m TxSubmitResult