Cardano.Benchmarking.Script.Action

Description

This is just exporting action, and liftToAction is tough to use because of the risk of circular imports.

Documentation

action :: Action -> ActionM ()

action has as its sole callers runScript from Cardano.Benchmark.Script and Selftest from Cardano.Benchmark.Script.Selftest. It translates the various cases of the Action to monadic values which execute the specified actions when evaluated. It passes all the cases' fields to functions with very similar names to the constructors.

liftToAction :: IO (Either TxGenError a) -> ActionM a

liftToAction first lifts from IO, then converts an Either to an ExceptT and then transforms the error type to Error.