Cardano.Benchmarking.Script.Env

Description

The Env type is the ADT for the state component of the ActionM monad stack. Its actual definition isn't exported in part because of a transition from an earlier very generic and polymorphic definition. In a number of respects, this module covers more of the ActionM like runActionM and liftTxGenError, but the only significant structure is Env for state. The accessors could likely be removed in favour of just using the record syntax to trim a few lines of code at the cost of exposing the structure's internals. Some of the naming related to the fact that Cardano.Benchmarking.Script.Action ran into circular dependency issues during the above transition.

Documentation

type ActionM a = ExceptT Error (RWST EnvConsts () Env IO) a

This abbreviates an ExceptT and RWST with particular types used as parameters.

data Env

The Env type represents the state maintained while executing a series of actions. The Maybe types are largely to represent as-of-yet unset values.

Constructors

Env

Fields

protoParams :: Maybe ProtocolParameterMode

ProtocolParameters is ultimately wrapped by ProtocolParameterMode which itself is a sort of custom Maybe.

envGenesis :: Maybe ShelleyGenesis

envProtocol :: Maybe SomeConsensusProtocol

envNetworkId :: Maybe NetworkId

envSocketPath :: Maybe FilePath

envKeys :: Map String (SigningKey PaymentKey)

envWallets :: Map String WalletRef

envSummary :: Maybe PlutusBudgetSummary

data Error where

Error adds two cases to TxGenError which in turn wraps Error implicit contexts to a couple of its constructors. These represent errors that might arise in the execution of a transaction with some distinctions as to the layers where the errors could arise. At this highest level, invalid users and wallets are potentially encountered. Plutus, protocol, API and some arbitrary errors are potentially encountered at the next layer. The layers correspond to Cardano.Benchmarking.Script.Core for the outermost and Cardano.Benchmarking.Set.Plutus for the middle, and the innermost to Cardano.Api.Error.

Constructors

TxGenError :: !TxGenError -> Error

UserError :: !String -> Error

WalletError :: !String -> Error

emptyEnv :: Env

Env uses Maybe to represent values that might be uninitialized. This being empty means Nothing is used across the board, along with all of the Map structures being empty.

newEnvConsts :: IOManager -> Maybe NixServiceOptions -> STM EnvConsts

runActionMEnv :: Env -> ActionM ret -> EnvConsts -> IO (Either Error ret, Env, ())

This runs an ActionM starting with the Env being passed.

liftTxGenError :: TxGenError -> ActionM a

This throws a Error in the ActionM monad.

liftIOSafe :: IO (Either TxGenError a) -> ActionM a

The safety comes from the invocation of throwE instead of just using the constructor for ExceptT to convert the Either to an ExceptT monadic value.

askIOManager :: ActionM IOManager

Accessor for the IOManager reader monad aspect of the RWST.

askNixSvcOpts :: ActionM (Maybe NixServiceOptions)

Accessor for the NixServiceOptions reader monad aspect of the RWST.

askEnvThreads :: ActionM (TVar (Maybe AsyncBenchmarkControl))

traceDebug :: String -> ActionM ()

Submit a debug message to the BenchTracers.

traceError :: String -> ActionM ()

Submit an error message to the BenchTracers.

traceBenchTxSubmit :: (forall txId. x -> TraceBenchTxSubmit txId) -> x -> ActionM ()

Helper to make submissions to the BenchTracers.

getBenchTracers :: ActionM BenchTracers

Read accessor for benchTracers. It would be burdensome on callers to have to have to case analyze this result. EnvConsts :: (Type -> Type) -> Type would make sense, using the pattern of data HKT f = HKT { f1 :: f t1, f2 :: f t2, ..} Then EnvConsts Maybe can be converted to EnvConsts Identity once initialization is complete so the main phase doesn't need to do this.

setBenchTracers :: BenchTracers -> ActionM ()

Write accessor for benchTracers.

getEnvGenesis :: ActionM ShelleyGenesis

Read accessor for envGenesis.

setEnvGenesis :: ShelleyGenesis -> ActionM ()

Write accessor for envGenesis.

getEnvKeys :: String -> ActionM (SigningKey PaymentKey)

Read accessor for envKeys.

setEnvKeys :: String -> SigningKey PaymentKey -> ActionM ()

Write accessor for envKeys.

getEnvNetworkId :: ActionM NetworkId

Read accessor for envNetworkId.

setEnvNetworkId :: NetworkId -> ActionM ()

Write accessor for envNetworkId.

getEnvProtocol :: ActionM SomeConsensusProtocol

Read accessor for envProtocol.

setEnvProtocol :: SomeConsensusProtocol -> ActionM ()

Write accessor for envProtocol.

getProtoParamMode :: ActionM ProtocolParameterMode

Read accessor for protoParams.

setProtoParamMode :: ProtocolParameterMode -> ActionM ()

Write accessor for protoParams.

getEnvSocketPath :: ActionM SocketPath

Read accessor for envSocketPath.

setEnvSocketPath :: FilePath -> ActionM ()

Write accessor for envSocketPath.

getEnvThreads :: ActionM (Maybe AsyncBenchmarkControl)

Read accessor for envThreads.

setEnvThreads :: AsyncBenchmarkControl -> ActionM ()

Write accessor for envThreads.

getEnvWallets :: String -> ActionM WalletRef

Read accessor for envWallets.

setEnvWallets :: String -> WalletRef -> ActionM ()

Write accessor for envWallets.

getEnvSummary :: ActionM (Maybe PlutusBudgetSummary)

Read accessor for envSummary.

setEnvSummary :: PlutusBudgetSummary -> ActionM ()

Write accessor for envSummary.