Cardano.Benchmarking.Script.Selftest

Description

The statically-defined data is the action list to execute, testScript. It actually does use a protocol file taken in from IO.

Documentation

runSelftest :: Env -> EnvConsts -> Maybe FilePath -> IO (Either Error ())

runSelftest is the interface to actually run the self-test. iom is the IO manager from Ouroboros.Network.IOManager. outFile is the file to output to; evalGenerator returns a transaction Stream that submitInEra does show and writeFile on.

printJSON :: IO ()

printJSON prints out the list of actions using Aeson. It has no callers within cardano-node.

testScript :: FilePath -> SubmitMode -> [Action]

testScript is a static list of Action parametrised with a file name and a mode indicating how to submit a transaction in SubmitMode passed along as a parameter within a Submit action.