Cardano.Benchmarking.Script.Types

Description

Action is likely the most impactful type exported, as it represents the individual steps to be executed by the transaction generator at the system level; however, Generator is much more of what one thinks of transactions themselves being. The Generator has to do with combining streams of the transactions one typically might think of doing with a wallet, where the Action level largely sees those transactions as interchangeable, and focuses more on the variety of things one might do with the connexion.

Documentation

data Action where

Action represents the individual actions to be executed by the tx-generator. It gets translated to ActionM using Env as the state, IOManager as the reader, and IO as the monad, and further wrapped in an ExceptT with an Error as the exception.

Constructors

SetNetworkId :: !NetworkId -> Action
SetNetworkId only entails changing a state variable in an Env.

SetSocketPath :: !FilePath -> Action
SetSocketPath likewise only entails a state variable change.

InitWallet :: !String -> Action
InitWallet just uses the name in a state variable and creates a fresh MVar with an empty FundQueue in it.

StartProtocol :: !FilePath -> !(Maybe FilePath) -> Action
StartProtocol sets state variables for protocol and genesis, but via mkNodeConfig and mkConsensusProtocol. The first unravels to reading wide a variety of config files in makeNodeConfiguration and The second unravels to reading genesis content in helper functions spread across eras.

Delay :: !Double -> Action
Delay translates to threadDelay via delay.

ReadSigningKey :: !String -> !(SigningKeyFile In) -> Action
ReadSigningKey translates to a readFileTextEnvelopeAnyOf from Cardano.Api.SerialiseTextEnvelope on the signing key file and then drops it into a state variable via setEnvKeys.

DefineSigningKey :: !String -> !(SigningKey PaymentKey) -> Action
DefineSigningKey is just a insert on the state variable.

AddFund :: !AnyCardanoEra -> !String -> !TxIn -> !Coin -> !String -> Action
AddFund is mostly a wrapper around walletRefInsertFund which in turn is just modifyMVar around insert.

WaitBenchmark :: Action
WaitBenchmark signifies a waitCatch on the AsyncBenchmarkControl for the environment and also folds tracers into the completion.

Submit :: !AnyCardanoEra -> !SubmitMode -> !TxGenTxParams -> !Generator -> Action
Submit mostly wraps benchmarkTxStream which in turn wraps walletBenchmark which in turn wraps txSubmissionClient and functions local to that like requestTxs.

CancelBenchmark :: Action
CancelBenchmark wraps a callback from the AsyncBenchmarkControl type, which is a shutdown action.

Reserved :: [String] -> Action
Reserved just emits an error and is a placeholder that helps with testing and quick fixes.

WaitForEra :: !AnyShelleyBasedEra -> Action

SetProtocolParameters :: ProtocolParametersSource -> Action
SetProtocolParameters has one option to read from a file and another to pass directly and just sets a state variable for the protoParams field of Env.

LogMsg :: !Text -> Action
LogMsg logs its message calling traceDebug i.e. via the tracer.

data Generator where

Generator is interpreted by evalGenerator as a series of transactions, albeit in the form of precursors to UTxO's.

Constructors

SecureGenesis :: !String -> !String -> !String -> Generator
SecureGenesis gets funds from a genesis via genesisSecureInitialFundForKey. This is where streams of transactions start.

Split :: !String -> !PayMode -> !PayMode -> [Coin] -> Generator
Split makes payments with change depending on the pay mode. The splitting is from potentially sending the change to a different place.

SplitN :: !String -> !PayMode -> !Int -> Generator
SplitN divides the funds by N and divides them up into that many transactions in a finite sequence. The handling starts from a case in evalGenerator and has some complexity to it.

NtoM :: !String -> !PayMode -> !NumberOfInputsPerTx -> !NumberOfOutputsPerTx -> !(Maybe Int) -> Maybe String -> Generator

Sequence :: [Generator] -> Generator
Sequence represents sequentially issuing a series in the form of a list of transaction series represented by Generator itself, but the nesting is done by first translating to ActionM, then a difficult-to-understand part follows in the form of for around each gList and is hard to interpret.

Cycle :: !Generator -> Generator
Cycle infinitely repeats the series of transactions in its argument, through evalGenerator as per cycle.

Take :: !Int -> !Generator -> Generator
Take is easily interpreted via take.

RoundRobin :: [Generator] -> Generator
RoundRobin wants interleaves but just errors out unimplemented.

OneOf :: [(Generator, Double)] -> Generator

data PayMode where

Constructors

PayToAddr :: !String -> !String -> PayMode

PayToScript :: !ScriptSpec -> !String -> PayMode

data ProtocolParameterMode where

Constructors

ProtocolParameterQuery :: ProtocolParameterMode

ProtocolParameterLocal :: ProtocolParameters -> ProtocolParameterMode

data ProtocolParametersSource where

Constructors

QueryLocalNode :: ProtocolParametersSource

UseLocalProtocolFile :: !FilePath -> ProtocolParametersSource

data ScriptBudget where

Constructors

StaticScriptBudget :: !FilePath -> !FilePath -> !ExecutionUnits -> !Bool -> ScriptBudget

AutoScript :: !FilePath -> !Int -> ScriptBudget

data ScriptSpec

Constructors

ScriptSpec

Fields

scriptSpecFile :: !(Either String FilePath)

scriptSpecBudget :: !ScriptBudget

scriptSpecPlutusType :: !TxGenPlutusType

data SubmitMode where

Constructors

LocalSocket :: SubmitMode

Benchmark :: !TargetNodes -> !TPSRate -> !NumberOfTxs -> SubmitMode

DumpToFile :: !FilePath -> SubmitMode

DiscardTX :: SubmitMode

NodeToNode :: NonEmpty NodeIPv4Address -> SubmitMode

type TargetNodes = NonEmpty NodeDescription

newtype TxList era

Constructors

TxList [Tx era]

Orphan instances

Eq (SigningKey PaymentKey)

Instance details

Methods

(==) :: SigningKey PaymentKey -> SigningKey PaymentKey -> Bool

(/=) :: SigningKey PaymentKey -> SigningKey PaymentKey -> Bool