Cardano.Benchmarking.Types

Description

Documentation

data SubmissionErrorPolicy

This controls how to react to errors during the benchmark.

Constructors

FailOnError

LogErrors

newtype ToAnnce tx

Transactions we decided to announce now.

Constructors

ToAnnce [tx]

newtype UnAcked tx

Transactions announced, yet unacked by peer.

Constructors

UnAcked [tx]

newtype Ack

Peer acknowledged this many txids of the outstanding window.

Constructors

Ack Int

newtype Req

Peer requested this many txids to add to the outstanding window.

Constructors

Req Int

newtype Sent

This many Txs sent to peer.

Constructors

Sent Int

newtype Unav

This many Txs requested by the peer, but not available for sending.

Constructors

Unav Int