Cardano.TxGenerator.Fund

Description

Types

Outputs that are available for spending. When building a new transaction, they provide the TxIn parts.

newtype Fund

InAnyCardanoEra helps form heterogenous collections wrt. eras.

Constructors

Fund

Fields

unFund :: InAnyCardanoEra FundInEra

data FundInEra era

FundInEra has field naming conventions suggesting anticipated use of lenses.

Constructors

FundInEra

Fields

_fundTxIn :: !TxIn

_fundWitness :: Witness WitCtxTxIn era

_fundVal :: !(TxOutValue era)

_fundSigningKey :: !(Maybe (SigningKey PaymentKey))

Accessors

Accessors for the various fields of FundInEra starting from a Fund.

getFundTxIn :: Fund -> TxIn

The TxIn is only a TxId and TxIx pair, and so era-independent.

getFundKey :: Fund -> Maybe (SigningKey PaymentKey)

Signing keys are optional as far as funds go.

getFundCoin :: Fund -> Coin

getFundWitness :: forall era. IsShelleyBasedEra era => Fund -> Witness WitCtxTxIn era

The Fund alternative is checked against cardanoEra, but getFundWitness otherwise wraps _fundWitness.