Cardano.TxGenerator.FundQueue

Description

Documentation

type FundQueue = Fifo Fund

A type alias for the specialized queue type.

toList :: FundQueue -> [Fund]

Converting to a list provides a different, non-FIFO access order.

emptyFundQueue :: FundQueue

This just restricts the type of the generic emptyFifo.

insertFund :: FundQueue -> Fund -> FundQueue

This just restricts the type of the generic insert function.

removeFund :: FundQueue -> Maybe (FundQueue, Fund)

This just restricts the type of the generic remove function.

removeFunds :: Int -> FundQueue -> Maybe (FundQueue, [Fund])

This just restricts the type of the generic removeN function.

removeAllFunds :: FundQueue -> (FundQueue, [Fund])

This is really just toList paired with an emptyFundQueue.

module Cardano.TxGenerator.Fund