Cardano.TxGenerator.Genesis

Description

This module provides means to secure funds that are given in genesis. To secure a fund, the key locking the transaction output in genesis has to be provided.

Documentation

genesisInitialFunds :: forall era. IsShelleyBasedEra era => NetworkId -> ShelleyGenesis -> [(AddressInEra era, Coin)]

genesisInitialFundForKey :: forall era. IsShelleyBasedEra era => NetworkId -> ShelleyGenesis -> SigningKey PaymentKey -> Maybe (AddressInEra era, Coin)

genesisTxInput :: NetworkId -> SigningKey PaymentKey -> TxIn

genesisExpenditure :: IsShelleyBasedEra era => NetworkId -> SigningKey PaymentKey -> AddressInEra era -> TxOutValue era -> Coin -> SlotNo -> SigningKey PaymentKey -> Either TxGenError (Tx era, Fund)

genesisSecureInitialFund :: forall era. IsShelleyBasedEra era => NetworkId -> ShelleyGenesis -> SigningKey PaymentKey -> SigningKey PaymentKey -> TxGenTxParams -> Either TxGenError (Tx era, Fund)

genesisValidate :: ShelleyGenesis -> Either String ()