Cardano.TxGenerator.PlutusContext

Documentation

data PlutusAutoLimitingFactor

Nothing references this type's name or uses its constructors outside this module.

Constructors

ExceededMemoryLimit

ExceededStepLimit

data PlutusBudgetFittingStrategy

This type specifies the end to which a script's loop counter is calibrated

Constructors

TargetTxExpenditure
calibrate for maximum expenditure of per-tx-budget

TargetBlockExpenditure (Maybe Double)
calibrate for maximum expenditure of per-block-budget, with a scaling factor of [1.0, 2.0]

TargetTxsPerBlock Int
calibrate for stable tx count per block

data PlutusBudgetSummary

This collects information describing the budget. It's only directly referenced in Cardano.Benchmarking.Script.Env to supply accessors as the state in an ActionM (RWST) monad and Cardano.Benchmarking.Script.Core where the only field touched is projectedTxSize.

Constructors

PlutusBudgetSummary

Fields

budgetPerBlock :: !ExecutionUnits

budgetPerTx :: !ExecutionUnits

budgetPerTxInput :: !ExecutionUnits

budgetStrategy :: !PlutusBudgetFittingStrategy

budgetTarget :: !ExecutionUnits

scriptId :: !FilePath

scriptArgDatum :: !ScriptData

scriptArgRedeemer :: !ScriptData

loopCounter :: !Int

loopLimitingFactors :: ![PlutusAutoLimitingFactor]

budgetUsedPerTxInput :: !ExecutionUnits

projectedBudgetUnusedPerBlock :: !ExecutionUnits

projectedBudgetUnusedPerTx :: !ExecutionUnits

projectedTxPerBlock :: !Int

projectedLoopsPerBlock :: !Int

projectedTxSize :: !(Maybe Int)

projectedTxFee :: !(Maybe Coin)

messageStrategy :: !(Maybe String)

human-readable description of fitting strategy used

messageId :: !String

one-string summary of fitting strategy and scaling factors used (only used by calibrate-script)

plutusAutoBudgetMaxOut :: ProtocolParameters -> ScriptInAnyLang -> PlutusAutoBudget -> PlutusBudgetFittingStrategy -> Int -> Either TxGenError (PlutusAutoBudget, Int, [PlutusAutoLimitingFactor])

Use a binary search to find a loop counter that maxes out the available script execution units. plutusAutoBudgetMaxOut makes two assumptions about the loop / PlutusAuto script: 1. The redeemer passed in is a valid one, and encodes i.a. the loop's termination value when counting down. 2. In the redeemer's argument structure, this value is the first numerical value that's encountered during traversal.

plutusAutoScaleBlockfit :: ProtocolParameters -> (FilePath, String) -> ScriptInAnyLang -> PlutusAutoBudget -> PlutusBudgetFittingStrategy -> Int -> Either TxGenError (PlutusBudgetSummary, PlutusAutoBudget, ExecutionUnits)

Can find the optimal scaling factor for block expenditure, by aiming at highest loop count per block iff TargetBlockExpenditure Nothing is given; will calibrate loop for any fully specified fitting strategy otherwise

plutusBudgetSummary :: ProtocolParameters -> (FilePath, String) -> PlutusBudgetFittingStrategy -> (PlutusAutoBudget, Int, [PlutusAutoLimitingFactor]) -> ExecutionUnits -> Int -> PlutusBudgetSummary

Only used in plutusAutoScaleBlockfit, this assembles a PlutusBudgetSummary from information about the how to budget. Some of the function arguments share names with the record fields mass imported with the Constr{..} notation, setting the field of the final result to that argument.

readScriptData :: FilePath -> IO (Either TxGenError HashableScriptData)

Load serialized ScriptData, filling in an empty value if no .json file is given

scriptDataModifyNumber :: (Integer -> Integer) -> ScriptData -> ScriptData

Modifies the first ScriptDataNumber encountered during traversal to the value provided

Orphan instances

FromJSON ScriptData

Instance details

Methods

parseJSON :: Value -> Parser ScriptData

parseJSONList :: Value -> Parser [ScriptData]

omittedField :: Maybe ScriptData

ToJSON ScriptData

Instance details

Methods

toJSON :: ScriptData -> Value

toEncoding :: ScriptData -> Encoding

toJSONList :: [ScriptData] -> Value

toEncodingList :: [ScriptData] -> Encoding

omitField :: ScriptData -> Bool