Cardano.TxGenerator.ProtocolParameters

The updatable protocol parameters

data ProtocolParameters

The values of the set of updatable protocol parameters. At any particular point on the chain there is a current set of parameters in use.

There are also parameters fixed in the Genesis file. See GenesisParameters.

Constructors

ProtocolParameters

Fields

protocolParamProtocolVersion :: (Natural, Natural)

Protocol version, major and minor. Updating the major version is used to trigger hard forks. (Major , Minor )

protocolParamDecentralization :: Maybe Rational

The decentralization parameter. This is fraction of slots that belong to the BFT overlay schedule, rather than the Praos schedule. So 1 means fully centralised, while 0 means fully decentralised.

This is the "d" parameter from the design document.

Deprecated in Babbage

protocolParamExtraPraosEntropy :: Maybe PraosNonce

Extra entropy for the Praos per-epoch nonce.

This can be used to add extra entropy during the decentralisation process. If the extra entropy can be demonstrated to be generated randomly then this method can be used to show that the initial federated operators did not subtly bias the initial schedule so that they retain undue influence after decentralisation.

protocolParamMaxBlockHeaderSize :: Natural

The maximum permitted size of a block header.

This must be at least as big as the largest legitimate block headers but should not be too much larger, to help prevent DoS attacks.

Caution: setting this to be smaller than legitimate block headers is a sure way to brick the system!

protocolParamMaxBlockBodySize :: Natural

The maximum permitted size of the block body (that is, the block payload, without the block header).

This should be picked with the Praos network delta security parameter in mind. Making this too large can severely weaken the Praos consensus properties.

Caution: setting this to be smaller than a transaction that can change the protocol parameters is a sure way to brick the system!

protocolParamMaxTxSize :: Natural

The maximum permitted size of a transaction.

Typically this should not be too high a fraction of the block size, otherwise wastage from block fragmentation becomes a problem, and the current implementation does not use any sophisticated box packing algorithm.

protocolParamTxFeeFixed :: Coin

The constant factor for the minimum fee calculation.

protocolParamTxFeePerByte :: Coin

Per byte linear factor for the minimum fee calculation.

protocolParamMinUTxOValue :: Maybe Coin

The minimum permitted value for new UTxO entries, ie for transaction outputs.

protocolParamStakeAddressDeposit :: Coin

The deposit required to register a stake address.

protocolParamStakePoolDeposit :: Coin

The deposit required to register a stake pool.

protocolParamMinPoolCost :: Coin

The minimum value that stake pools are permitted to declare for their cost parameter.

protocolParamPoolRetireMaxEpoch :: EpochInterval

The maximum number of epochs into the future that stake pools are permitted to schedule a retirement.

protocolParamStakePoolTargetNum :: Word16

The equilibrium target number of stake pools.

This is the "k" incentives parameter from the design document.

protocolParamPoolPledgeInfluence :: Rational

The influence of the pledge in stake pool rewards.

This is the "a_0" incentives parameter from the design document.

protocolParamMonetaryExpansion :: Rational

The monetary expansion rate. This determines the fraction of the reserves that are added to the fee pot each epoch.

This is the "rho" incentives parameter from the design document.

protocolParamTreasuryCut :: Rational

The fraction of the fee pot each epoch that goes to the treasury.

This is the "tau" incentives parameter from the design document.

protocolParamCostModels :: Map AnyPlutusScriptVersion CostModel

Cost models for script languages that use them.

Introduced in Alonzo

protocolParamPrices :: Maybe ExecutionUnitPrices

Price of execution units for script languages that use them.

Introduced in Alonzo

protocolParamMaxTxExUnits :: Maybe ExecutionUnits

Max total script execution resources units allowed per tx

Introduced in Alonzo

protocolParamMaxBlockExUnits :: Maybe ExecutionUnits

Max total script execution resources units allowed per block

Introduced in Alonzo

protocolParamMaxValueSize :: Maybe Natural

Max size of a Value in a tx output.

Introduced in Alonzo

protocolParamCollateralPercent :: Maybe Natural

The percentage of the script contribution to the txfee that must be provided as collateral inputs when including Plutus scripts.

Introduced in Alonzo

protocolParamMaxCollateralInputs :: Maybe Natural

The maximum number of collateral inputs allowed in a transaction.

Introduced in Alonzo

protocolParamUTxOCostPerByte :: Maybe Coin

Cost in ada per byte of UTxO storage.

Introduced in Babbage

newtype LedgerProtocolParameters era

Constructors

LedgerProtocolParameters

Fields

unLedgerProtocolParameters :: PParams (ShelleyLedgerEra era)

convertToLedgerProtocolParameters :: ShelleyBasedEra era -> ProtocolParameters -> Either ProtocolParametersConversionError (LedgerProtocolParameters era)

toLedgerPParams :: ShelleyBasedEra era -> ProtocolParameters -> Either ProtocolParametersConversionError (PParams (ShelleyLedgerEra era))

fromLedgerPParams :: ShelleyBasedEra era -> PParams (ShelleyLedgerEra era) -> ProtocolParameters