Cardano.TxGenerator.Script.Types

Description

this module is currently experimental and serves as testing ground for a new script design for the generator.

Documentation

type TxGenScript = [RootNode]

data RootNode

Constructors

SigningKeys [NamedKey]

Wallets

Params TxGenTxParams

Script

data NamedKey

Constructors

KeyLiteral

Fields

name :: Text

base16Cbor :: Text

KeyFile

Fields

name :: Text

filePath :: FilePath

jsonOptions :: Options

testRender :: IO ()

Orphan instances

ToJSON TxGenTxParams

Instance details

Methods

toJSON :: TxGenTxParams -> Value

toEncoding :: TxGenTxParams -> Encoding

toJSONList :: [TxGenTxParams] -> Value

toEncodingList :: [TxGenTxParams] -> Encoding

omitField :: TxGenTxParams -> Bool

Generic TxGenTxParams

Instance details

Associated Types

type Rep TxGenTxParams :: Type -> Type

Methods

from :: TxGenTxParams -> Rep TxGenTxParams x

to :: Rep TxGenTxParams x -> TxGenTxParams