Cardano.TxGenerator.Setup.NodeConfig

Description

This module provides convenience functions when dealing with cardano-node's configuration and config files as well as protocol instantiation.

Documentation

mkConsensusProtocol :: NodeConfiguration -> IO (Either TxGenError SomeConsensusProtocol)

getGenesis :: SomeConsensusProtocol -> ShelleyGenesis

extract genesis from a Cardano protocol NB. this helper is *only* for protocols created with this module as this guarantees proper error handling when trying to create a non-Cardano protocol.

getGenesisPath :: NodeConfiguration -> Maybe GenesisFile

extract the path to genesis file from a NodeConfiguration for Cardano protocol

mkNodeConfig :: FilePath -> IO (Either TxGenError NodeConfiguration)

Creates a NodeConfiguration from a config file; the result is devoid of any keys/credentials