Cardano.TxGenerator.Setup.SigningKey

Description

This module provides convenience functions when dealing with signing keys.

Documentation

parseSigningKeyTE :: TextEnvelope -> Either TxGenError (SigningKey PaymentKey)

parseSigningKeyBase16 :: ByteString -> Either TxGenError (SigningKey PaymentKey)

readSigningKeyFile :: SigningKeyFile 'In -> IO (Either TxGenError (SigningKey PaymentKey))

data PaymentKey

Shelley-era payment keys. Used for Shelley payment addresses and witnessing transactions that spend from these addresses.

This is a type level tag, used with other interfaces like Key.

data family SigningKey keyrole

The type of cryptographic signing key, for each key role.