Cardano.TxGenerator.Types

Description

Documentation

type ShelleyGenesis = ShelleyGenesis

data TxGenPlutusParams

Constructors

PlutusOn
Generate Plutus Txs for given script

Fields

plutusType :: !TxGenPlutusType

plutusScript :: !(Either String FilePath)

name or path of the Plutus script

plutusDatum :: !(Maybe FilePath)

Datum passed to the Plutus script (JSON file in ScriptData schema)

plutusRedeemer :: !(Maybe FilePath)

Redeemer passed to the Plutus script (JSON file in ScriptData schema)

plutusExecMemory :: !(Maybe Natural)

Max. memory for ExecutionUnits (overriding corresponding protocol parameter)

plutusExecSteps :: !(Maybe Natural)

Max. steps for ExecutionUnits (overriding corresponding protocol parameter)

PlutusOff
Do not generate Plutus Txs

type NumberOfTxs = Int

type TPSRate = Double

data TxGenError where

Constructors

ApiError :: Error e => !e -> TxGenError

ProtocolError :: Error e => !e -> TxGenError

PlutusError :: Show e => !e -> TxGenError

TxGenError :: !String -> TxGenError

data TxEnvironment era

Constructors

TxEnvironment

Fields

txEnvNetworkId :: !NetworkId

txEnvProtocolParams :: !ProtocolParameters

txEnvFee :: TxFee era

txEnvMetadata :: TxMetadataInEra era

type TxGenerator era = [Fund] -> [TxOut CtxTx era] -> Either TxGenError (Tx era, TxId)

data TxGenPlutusResolvedTo

Documents how the plutusScript parameter above was eventually resolved

Constructors

ResolvedToLibrary String
source is the library from the plutus-scripts-bench package

ResolvedToFallback FilePath
source it the tx-generator's scripts-fallback data directory

ResolvedToFileName FilePath
source is a .plutus file

type NumberOfInputsPerTx = Int

type NumberOfOutputsPerTx = Int

type TxAdditionalSize = Int

type FundSource m = m (Either TxGenError [Fund])

type FundToStore m = Fund -> m ()

type FundToStoreList m = [Fund] -> m ()

data PayWithChange

Constructors

PayExact [Coin]

PayWithChange Coin [Coin]

data TxGenTxParams

Constructors

TxGenTxParams

Fields

txParamFee :: !Coin

Transaction fee, in Lovelace

txParamAddTxSize :: !Int

Extra transaction payload, in bytes -- Note [Tx additional size]

txParamTTL :: !SlotNo

Time-to-live

data TxGenConfig

Constructors

TxGenConfig

Fields

confMinUtxoValue :: !Coin

Minimum value required per UTxO entry

confTxsPerSecond :: !Double

Strength of generated workload, in transactions per second

confInitCooldown :: !Double

Delay between init and main submissions in seconds

confTxsInputs :: !NumberOfInputsPerTx

Inputs per transaction

confTxsOutputs :: !NumberOfOutputsPerTx

Outputs per transaction

data TxGenPlutusType

Constructors

LimitSaturationLoop
Generate Txs for a Plutus loop script, choosing settings to max out per Tx budget

LimitTxPerBlock_8
Generate Txs for a Plutus loop script, choosing settings to best fit 8 Txs into block budget

LimitTxPerBlock_4
Generate Txs for a Plutus loop script, choosing settings to best fit 4 Txs into block budget

BenchCustomCall
Built-in script for benchmarking various complexity of data passed via Plutus API

CustomScript

data PlutusAutoBudget

Constructors

PlutusAutoBudget
Specifies a budget and parameters for a PlutusAuto loop script

Fields

autoBudgetUnits :: !ExecutionUnits

execution units available per Tx input / script run

autoBudgetDatum :: !ScriptData

datum for the auto script

autoBudgetRedeemer :: !ScriptRedeemer

valid redeemer for the auto script

autoBudgetUpperBoundHint :: !(Maybe Int)

hints at a loop count upper bount; speeds up calibration for scripts with low loop counts, but does not influence outcome

defaultTxGenTxParams :: TxGenTxParams

isPlutusMode :: TxGenPlutusParams -> Bool

hasLoopCalibration :: TxGenPlutusType -> Bool

hasStaticBudget :: TxGenPlutusParams -> Maybe ExecutionUnits