Cardano.TxGenerator.Utils

Documentation

mkTxIn :: HasCallStack => Text -> TxIn

Convert text representation of a txin "hash#txid" to a TxIn e.g. "dbaff4e270cfb55612d9e2ac4658a27c79da4a5271c6f90853042d1403733810#0" Partial. Useful in tests.

inputsToOutputsWithFee :: Coin -> Int -> [Coin] -> [Coin]

liftAnyEra :: (forall era. IsCardanoEra era => f1 era -> f2 era) -> InAnyCardanoEra f1 -> InAnyCardanoEra f2

liftAnyEra applies a function to the value in InAnyCardanoEra regardless of which particular era.

keyAddress :: forall era. IsShelleyBasedEra era => NetworkId -> SigningKey PaymentKey -> AddressInEra era

keyAddress determines an address for the relevant era.

includeChange :: Coin -> [Coin] -> [Coin] -> PayWithChange

includeChange gets use made of it as a value splitter in sourceToStoreTransactionNew by evalGenerator.

mkTxFee :: IsShelleyBasedEra era => Coin -> TxFee era

mkTxFee reinterprets the Either returned by txFeesExplicitInEra with TxFee constructors.

mkTxValidityUpperBound :: forall era. IsShelleyBasedEra era => SlotNo -> TxValidityUpperBound era

mkTxValidityUpperBound rules out needing the TxValidityNoUpperBound with the constraint of IsShelleyBasedEra.

mkTxInModeCardano :: IsShelleyBasedEra era => Tx era -> TxInMode

mkTxInModeCardano never uses the TxInByronSpecial constructor because its type enforces it being a Shelley-based era.