Cardano.Benchmarking.OuroborosImports

Documentation

type CardanoBlock = CardanoBlock StandardCrypto

type LocalSubmitTx = TxInMode -> IO (SubmitResult TxValidationErrorInCardanoMode)

data LoggingLayer

The LoggingLayer interface that we can expose. We want to do this since we want to be able to mock out any function tied to logging.

The good side of this is that _each function has it's own effects_ and that is ideal for tracking the functions effects and constraining the user (programmer) of those function to use specific effects in them. https://github.com/input-output-hk/cardano-sl/blob/develop/util/src/Pos/Util/Log/LogSafe.hs

data PaymentKey

Shelley-era payment keys. Used for Shelley payment addresses and witnessing transactions that spend from these addresses.

This is a type level tag, used with other interfaces like Key.

data ShelleyGenesis c

Shelley genesis information

Note that this is needed only for a pure Shelley network, hence it being defined here rather than in its own module. In mainnet, Shelley will transition naturally from Byron, and thus will never have its own genesis information.

data family SigningKey keyrole

The type of cryptographic signing key, for each key role.

type SigningKeyFile = File (SigningKey ())

type StandardShelley = ShelleyEra StandardCrypto

The Shelley era with standard crypto

data NetworkId

makeLocalConnectInfo :: NetworkId -> SocketPath -> LocalNodeConnectInfo

protocolToTopLevelConfig :: SomeConsensusProtocol -> TopLevelConfig CardanoBlock

protocolToNetworkId :: SomeConsensusProtocol -> NetworkId

protocolToCodecConfig :: SomeConsensusProtocol -> CodecConfig CardanoBlock

submitTxToNodeLocal :: MonadIO m => LocalNodeConnectInfo -> TxInMode -> m (SubmitResult TxValidationErrorInCardanoMode)