Cardano.TxGenerator.ProtocolParameters

The updatable protocol parameters

data ProtocolParameters

The values of the set of updatable protocol parameters. At any particular point on the chain there is a current set of parameters in use.

There are also parameters fixed in the Genesis file. See GenesisParameters.

Constructors

ProtocolParameters (Natural, Natural) (Maybe Rational) (Maybe PraosNonce) Natural Natural Natural Coin Coin (Maybe Coin) Coin Coin Coin EpochInterval Word16 Rational Rational Rational (Map AnyPlutusScriptVersion CostModel) (Maybe ExecutionUnitPrices) (Maybe ExecutionUnits) (Maybe ExecutionUnits) (Maybe Natural) (Maybe Natural) (Maybe Natural) (Maybe Coin)

convertToLedgerProtocolParameters :: ShelleyBasedEra era -> ProtocolParameters -> Either ProtocolParametersConversionError (LedgerProtocolParameters era)

toLedgerPParams :: ShelleyBasedEra era -> ProtocolParameters -> Either ProtocolParametersConversionError (PParams (ShelleyLedgerEra era))

fromLedgerPParams :: ShelleyBasedEra era -> PParams (ShelleyLedgerEra era) -> ProtocolParameters